// sitat-fe-workspace/projects/app-programmi/src/app/modules/models/interventi/interventi.model.ts
export interface InterventoEntry {
    acqVerdi?: number;
    acquistoBeniRiciclati?: number;
    acquistoMaterialiRiciclati?: number;
    acquistoRicompreso?: number;
    altreRisorseDisp1?: number;
    altreRisorseDisp2?: number;
    altreRisorseDisp3?: number;
    altreRisorseDispSucc?: number;
    annoAvvio?: number;
    annoInsAcquisto?: number;
    avcpv?: string;
    categoriaIntervento?: string;
    classificazioneIntervento?: string;
    codCpv?: string;
    codIstatComune?: string;
    codiceAusa?: string;
    codiceCup?: string;
    codiceInterno?: string;
    codiceNuts?: string;
    codiceRup?: string;
    coperturaFinanziaria?: string;
    cuiCollegato?: string;
    delegaProcAff?: string;
    descrizione?: string;
    direzioneGenerale?: string;
    dirigenteUfficio?: string;
    durataContratto?: number;
    esenteCup?: string;
    finalitaIntervento?: string;
    id?: number;
    idProgramma?: number;
    immobili?: Array<ImmobileEntry>;
    impDestVincolo1?: number;
    impDestVincolo2?: number;
    impDestVincolo3?: number;
    impDestVincoloSucc?: number;
    importoAcqMututo1?: number;
    importoAcqMututo2?: number;
    importoAcqMututo3?: number;
    importoAcqMututoSucc?: number;
    importoAlnettoIvaAv?: number;
    importoAlnettoIvaMr?: number;
    importoCapPriv1?: number;
    importoCapPriv2?: number;
    importoCapPriv3?: number;
    importoCapPrivSucc?: number;
    importoDispPriv1?: number;
    importoDispPriv2?: number;
    importoDispPriv3?: number;
    importoDispPrivSucc?: number;
    importoFinanz1?: number;
    importoFinanz2?: number;
    importoFinanz3?: number;
    importoFinanzSucc?: number;
    importoIva1?: number;
    importoIva2?: number;
    importoIva3?: number;
    importoIvaAv?: number;
    importoIvaMr?: number;
    importoIvaSucc?: number;
    importoRisorseFinanz?: number;
    importoRisorseFinanzAltro?: number;
    importoRisorseRegionali?: number;
    importoTotale?: number;
    importoTotaleAv?: number;
    importoTotaleMr?: number;
    inPianoAnnuale?: string;
    lavoroComplesso?: string;
    livelloPriorita?: number;
    lottoFunzionale?: string;
    meseAvvioProc?: number;
    mrcpv?: string;
    normativaRiferimento?: string;
    note?: string;
    numProgetto?: number;
    numeroCui?: string;
    nuovoAffidamento?: string;
    oggettoAv?: string;
    oggettoMr?: string;
    proceduraAffidamento?: number;
    quantita?: number;
    referenteDatiComunicati?: string;
    risorseDicapitolo?: Array<RisorsaCapitoloBaseEntry>;
    scadenzaUtilizzoFinanziamento?: Date;
    settore?: string;
    soggettoDelegato?: string;
    speseSostenute?: number;
    stanziamentiBilancio1?: number;
    stanziamentiBilancio2?: number;
    stanziamentiBilancio3?: number;
    stanziamentiBilancioSucc?: number;
    statoProgettazioneApprovata?: string;
    strutturaOperativa?: string;
    tipologiaCapitalePrivato?: string;
    tipologiaIntervento?: number;
    totImpDispFin?: number;
    totaleCapitalePriv?: number;
    trasfImmobili1?: number;
    trasfImmobili2?: number;
    trasfImmobili3?: number;
    trasfImmobiliSucc?: number;
    unitaMisura?: number;
    valutazioneResp?: string;
    variato?: number;    
    cigAccordoQuadro?: string;
    verificaConfAmbiente?: string;
    verificaConfUrbanistica?: string;
}


export interface RisorsaCapitoloBaseEntry {
    id?: number;
    idIntervento?: number;
    idProgramma?: number;
    importoComplessivo1?: number;
    importoComplessivo2?: number;
    importoComplessivo3?: number;
    importoComplessivoSucc?: number;
}